+++ net-gateway/src/endpoints/charts/network_bandwidth_per_endpoint/endpoint.rs
@@ -12,6 +12,7 @@ use crate::authorization::mock_authenticator::MockAuthenticator;
 use crate::core::app_state::AppState;
 use crate::core::client_data::ClientData;
 use crate::core::filter::FiltersWrapper;
+use crate::core::filter::FiltersWrapper;
 use crate::core::general_filters::GeneralFilters;
 
 use crate::core::service_request_management::service_request_manager::ServiceRequestManager;
@@ -24,6 +25,7 @@ async fn get_bandwidth_per_endpoint(
     state: web::Data<AppState>,
     client_data: web::Query<ClientData>,
     params: web::Query<GeneralFilters>,
+    filters_wrapper: web::Query<FiltersWrapper>,
     filters_wrapper: web::Query<FiltersWrapper>,
     req: HttpRequest,
 ) -> impl Responder {